cloud-log-platform/log-processor/requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
prometheus-client==0.19.0
python-dotenv==1.0.0
